requirements.txt
dash>=2.14
plotly>=5.18
pandas>=2.1
garminconnect>=0.2
python-dotenv>=1.0
